refactor 03.12.25

+++ app/Http/Controllers/JobController.php
@@ -6,10 +6,16 @@ use App\Models\Job;
 use App\Models\Tag;
 use App\Http\Requests\StoreJobRequest;
 use App\Http\Requests\UpdateJobRequest;
+use App\Models\Employer;
+use App\Models\User;
+#use Illuminate\Auth\Access\Gate;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+
 
 class JobController extends Controller
 {
+    
     /**
      * Display a listing of the resource.
     */
@@ -33,6 +39,7 @@ class JobController extends Controller
     {
         return view('jobs.create', [
             'tags' => Tag::all(),
+            'employers' => Employer::all(),
         ]);
     }
 
@@ -41,7 +48,7 @@ class JobController extends Controller
      */
     public function store(StoreJobRequest $request)
     {
-         request()->validate([
+        request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required']
         ]);
@@ -78,6 +85,21 @@ class JobController extends Controller
     public function edit(Job $job)
     { 
         #$job->load('tags');
+        #Auth::user()->can('edit', $job);
+
+        // Gate::define('edit-job', function(User $user,Job $job){
+        //     return $job->employer->user->is($user());
+        // });
+
+        // if(Auth::guest()){
+        //     return redirect('/login');
+        // }
+
+        #Gate::authorize('edit-job', $job);
+
+        // if($job->employer->user->isNot(Auth::user())){
+        // abort(403, 'Unauthorized action.'); 
+        // }
 
         return view('jobs.edit', [
             'job' => $job
@@ -89,6 +111,8 @@ class JobController extends Controller
      */
     public function update(UpdateJobRequest $request, Job $job)
     {
+        Gate::authorize('edit-job', $job);
+
         $job->update($request->validated());
         return redirect()->route('jobs.show', $job)->with('success', 'Job updated successfully.');
     }
@@ -98,6 +122,8 @@ class JobController extends Controller
      */
     public function destroy(Job $job)
     {
+        Gate::authorize('edit-job', $job);
+
         $job->delete();
         return redirect()->route('jobs.index')->with('success', 'Job deleted successfully.');
     }
